+++ kinto/plugins/quotas/__init__.py
@@ -5,7 +5,8 @@ from .listener import on_resource_changed
 
 def includeme(config):
     config.add_api_capability('quotas',
-                              description='Quotas Management on Buckets.',
+                              description='Quotas Management on buckets '
+                                          'and collections.',
                               url='https://kinto.readthedocs.io')
 
     # Listen to every resources (except history)