+++ CHANGELOG.rst
@@ -27,7 +27,7 @@ This document describes changes between each past release.
   setting. Useful for Datadog™ integration for example (fixes #626).
 - Added a ``delete-collection`` action to the ``kinto`` command. (#727)
 - Added verbosity options to the ``kinto`` command. (#745)
-- Added a quotas plugin. (#752)
+- Added a built-in plugin that allows to define quotas per bucket or collection. (#752)
 
 **Bug fixes**